pyproject.toml
[project]
name = "scheduler-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.167.0",
    "google-auth>=2.39.0",
    "google-auth-oauthlib>=1.2.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
]
